requirements.txt
flask[async]
python-dotenv
requests
gunicorn